mlb-predictor/requirements.txt
# ── Core ──────────────────────────────────────────────────
pandas>=2.2,<3
numpy>=2.1,<3
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2

# ── Data sources ──────────────────────────────────────────
pybaseball>=2.2.7,<3
requests>=2.31,<3

# ── ML – v1 baselines ────────────────────────────────────
scikit-learn>=1.5,<2

# ── ML – v2 boosting (install when ready) ─────────────────
# xgboost>=2.0,<3
# lightgbm>=4.3,<5

# ── Feature snapshots & backtest artifacts ────────────────
pyarrow>=18,<21

# ── API (Phase 4) ────────────────────────────────────────
fastapi>=0.111,<1
uvicorn[standard]>=0.29,<1

# ── Dev / test ────────────────────────────────────────────
pytest>=8,<9
black>=24,<25
isort>=5.13,<6

